Tiny minor amendsments to the gitignore file, and IMPORTANT-README Notes Just a few silly things I forgot to do.
Also I realised that pushing to a git repo with a working tree doesn't work
because Git is cowardly and wont clobber the files in the working tree ;-) Nice work Linus
So I changed the Repo structure on the Production server. Creating a --bare repository in
/home/ricktimmis/sparkcrm.git
I then created some scripts to do the code deployment into production. I know it adds another layer
of complexity, but it's either right or it's wrong.